Unified graph correction methods and made them optional
