+++ src/algorithms/mod.rs
@@ -1,5 +1,5 @@
 //! Various algorithms used to create, modify and collapse graphs
-pub mod pruner;
+pub mod corrector;
 pub mod collapser;
 pub mod builder;
 pub mod shrinker;
