+++ src/algorithms/shrinker.rs
@@ -1,11 +1,10 @@
 //! Shrink the given graph
 use crate::graphs::Graph;
-use crate::graphs::pt_graph::{CarryType, EdgeIndex, EdgeType, EmptyEdge, NodeIndex, NodeStruct, PtGraph};
+use crate::graphs::pt_graph::{CarryType, EdgeIndex, EdgeType, EmptyEdge, NodeIndex, PtGraph};
 use crate::logw;
 
 use petgraph::Direction::{Incoming, Outgoing};
 use petgraph::visit::EdgeRef;
-use std::{cmp::{min, max}, collections::BTreeSet};
 
 // use std::process::exit;
 
@@ -33,208 +32,6 @@ pub trait Shrinkable {
 
 }
 
-/// Checks whether the candidate area can be a good bubble for error correction.
-pub fn check_bubble_structure(ptgraph: &PtGraph, startn : NodeIndex, invec : Vec<EdgeIndex>) -> bool {
-    let mut midnodes = Vec::with_capacity(2);
-    let mut midcts   = Vec::with_capacity(2);
-    let outnode : NodeIndex;
-    for e in invec {
-        midnodes.push(ptgraph.graph.edge_endpoints(e).unwrap().1);
-        midcts.push(ptgraph.graph.edge_weight(e).unwrap().t.get_from_and_to().1);
-    }
-
-    // We need to check that the two intermediate nodes are different
-    if midnodes[0] == midnodes[1] || midnodes[0] == startn || midnodes[1] == startn {
-        return false;
-    }
-
-    // Now, how many neighbours do we have from the middle nodes? It should be only one, and the same
-    // and also, only one incoming neighbour!
-    let tmpv0 = ptgraph.out_neighbours_bi(midnodes[0], midcts[0]);
-    if tmpv0.len() != 1 || ptgraph.in_neighbours_bi(midnodes[0], midcts[0]).len() != 1 {
-        return false
-    }
-    outnode = tmpv0[0].0;
-    let tmpv1 = ptgraph.out_neighbours_bi(midnodes[1], midcts[1]);
-    let outct = tmpv0[0].1.get_from_and_to().1;
-
-    if (tmpv1.len() != 1) || (tmpv1[0].0 != outnode) || (outct != tmpv1[0].1.get_from_and_to().1) || (ptgraph.in_neighbours_bi(midnodes[1], midcts[1]).len() != 1) {
-        return false;
-    }
-
-    // and now we check that we go to a new node that is different from outnode (i.e. no self-edges/loops)
-    let tmpv3 = ptgraph.out_neighbours_bi(outnode, outct);
-    if tmpv3.len() != 1 || tmpv3[0].0 == startn || ptgraph.in_neighbours_bi(outnode, outct).len() != 2 {
-        return false;
-    }
-
-    return true;
-}
-
-
-/// This function collapses standard bubbles depending on the number of counts (very naive)
-pub fn collapse_bubble(ptgraph: &mut PtGraph, startn : NodeIndex) -> bool {
-    // Remember: we always start with the Min being the origin of the two outward edges
-    // let prevconn = ptgraph.in_neighbours_min(startn)[0];
-    let midconns = ptgraph.out_neighbours_min(startn);
-    if midconns.len() != 2 || ptgraph.out_neighbours_bi(midconns[0].0, midconns[0].1.get_from_and_to().1).len() != 1 {
-        return false;
-    }
-    let chosennode : usize;
-    let node0w = ptgraph.graph.node_weight(midconns[0].0).unwrap();
-    let node1w = ptgraph.graph.node_weight(midconns[1].0).unwrap();
-    let savedmidw : NodeStruct;
-
-    let count_threshold = min((0.1 * max(node0w.counts, node1w.counts) as f32).round() as u16, 1); // Inspired by Skesa
-
-    // println!("here1");
-
-    if node0w.counts < count_threshold {
-        chosennode = 1;
-        savedmidw = node1w.clone();
-        // println!("here3");
-    } else if node1w.counts < count_threshold {
-        chosennode = 0;
-        savedmidw = node0w.clone();
-        // println!("here4");
-    } else {
-        // println!("here2");
-        // We have not managed to exclude one of the two options. What do we do now?
-
-        if ((node0w.abs_ind.len() - node1w.abs_ind.len()) as i32).abs() as f32/(max(node0w.abs_ind.len(), node1w.abs_ind.len()) as f32) > 0.025 {
-            // There is a large difference in the number of k-mers between the options: this could be a coincidence link between
-            // two locations far away in the genome. Let's compare with the counts of the start and end nodes.
-            // println!("here5");
-
-            let startn_counts = ptgraph.graph.node_weight(startn).unwrap().counts;
-
-            // println!("preasdf");
-            let endn_counts   = ptgraph.graph.node_weight(ptgraph.out_neighbours_bi(midconns[0].0, midconns[0].1.get_from_and_to().1)[0].0).unwrap().counts;
-            // println!("asdfadsf");
-            let average_surrounding_counts = ((startn_counts + endn_counts) as f32/2.0).round() as u16;
-
-            let rel_diff_0 = ((node0w.counts - average_surrounding_counts) as i32).abs() as f32/(average_surrounding_counts as f32);
-            let rel_diff_1 = ((node1w.counts - average_surrounding_counts) as i32).abs() as f32/(average_surrounding_counts as f32);
-
-
-            // NOTE: in the future, removing valid connections between nodes might not be desirable, as we might be able to resolve these
-            // unitigs with e.g. larger k-value graphs.
-            if        rel_diff_0 >  0.2 && rel_diff_1 <= 0.2 {
-                // 1 has similar counts as the neighbours
-                // We remove the connections with 0
-                // println!("here6");
-                let tmpvec : Vec<EdgeIndex> = ptgraph.graph.edges_directed(midconns[0].0, Outgoing).chain(ptgraph.graph.edges_directed(midconns[0].0, Incoming)).map(|er| er.id()).collect();
-                for e in tmpvec {
-                    ptgraph.graph.remove_edge(e);
-                }
-
-            } else if rel_diff_0 <= 0.2 && rel_diff_1 >  0.2 {
-                // 0 has similar counts as the neighbours
-                // We remove the connections with 1
-                // println!("here7");
-                let tmpvec : Vec<EdgeIndex> = ptgraph.graph.edges_directed(midconns[1].0, Outgoing).chain(ptgraph.graph.edges_directed(midconns[1].0, Incoming)).map(|er| er.id()).collect();
-                for e in tmpvec {
-                    ptgraph.graph.remove_edge(e);
-                }
-
-            } else {
-                // println!("here8");
-                // Both have different counts as the neighbours
-                // Perhaps both connections are unlikely. We remove all the connections of the bubble (i.e. we create a contig break).
-                let tmpvec : Vec<EdgeIndex> = ptgraph.graph.edges_directed(midconns[0].0, Outgoing).chain(ptgraph.graph.edges_directed(midconns[0].0, Incoming)).chain(ptgraph.graph.edges_directed(midconns[1].0, Outgoing).chain(ptgraph.graph.edges_directed(midconns[1].0, Incoming))).map(|er| er.id()).collect();
-                for e in tmpvec {
-                    ptgraph.graph.remove_edge(e);
-                }
-            }
-            // In these cases we exit as there is nothing else to do
-            return true;
-
-        } else {
-            // println!("here9");
-            // The difference in the number of k-mers is very small, this could be a SNP. Let's pick up the highest count one.
-            if node0w.counts > node1w.counts {
-                // println!("here10");
-                chosennode = 0;
-                savedmidw = node0w.clone();
-            } else if node0w.counts < node1w.counts {
-                // println!("here11");
-                chosennode = 1;
-                savedmidw = node1w.clone();
-            } else if node0w.abs_ind.len() > node1w.abs_ind.len() {
-                // println!("here12");
-                chosennode = 0;
-                savedmidw = node0w.clone();
-            } else {
-                // println!("here13");
-                chosennode = 1;
-                savedmidw = node1w.clone();
-            }
-        }
-    }
-
-
-    // // Decision time! (simple)
-    // if node0w.counts > node1w.counts {
-    //     chosennode = 0;
-    //     savedmidw = node0w.clone();
-    // } else if node0w.counts < node1w.counts {
-    //     chosennode = 1;
-    //     savedmidw = node1w.clone();
-    // } else if node0w.abs_ind.len() > node1w.abs_ind.len() {
-    //     chosennode = 0;
-    //     savedmidw = node0w.clone();
-    // } else {
-    //     chosennode = 1;
-    //     savedmidw = node1w.clone();
-    // }
-
-    // We need to know if we are arriving, at the exit node of the bubble, to the Max ct.
-    let midnodect = midconns[chosennode].1.get_from_and_to().1;
-    let midconn2  = ptgraph.out_neighbours_bi(midconns[chosennode].0, midnodect)[0];
-    let outct     = midconn2.1.get_from_and_to().1;
-    let outconn   = ptgraph.out_neighbours_bi(midconn2.0, outct)[0];
-    let savedoutw = ptgraph.graph.node_weight(midconn2.0).unwrap().clone();
-
-    // By construction, the start node and the end node MUST have only one k-mer. If not, something is wrong.
-    if ptgraph.graph.node_weight(startn).unwrap().abs_ind.len() > 1 || ptgraph.graph.node_weight(midconn2.0).unwrap().abs_ind.len() > 1 {
-        panic!("Trying to remove a bubble with a start or end node with more than one k-mer!");
-    };
-
-    // We have copied the weights and have the information we need to continue, we can thus erase all the
-    // nodes except from the start node.
-    ptgraph.graph.remove_node(midconns[0].0); // intermediate node 0
-    ptgraph.graph.remove_node(midconns[1].0); // intermediate node 1
-    ptgraph.graph.remove_node(midconn2.0);    // end node
-    // At the end, the final node (the original start node) will have an internal node with and edge
-    // MinToMin, as we are starting from the Min ct in any case. However, we might need to adjust the
-    // following edges from the end node if we exit through the Max one.
-
-    let mutrefw = ptgraph.graph.node_weight_mut(startn).unwrap();
-
-    mutrefw.merge(&savedmidw, midconns[chosennode].1); // Intermediate node k-mer(s)
-    mutrefw.set_internal_edge(EdgeType::MinToMin);     // Inner edge
-    mutrefw.abs_ind.push(savedoutw.abs_ind[0]);        // Out node k-mer
-    mutrefw.set_mean_counts(&vec![mutrefw.counts, savedmidw.counts, savedoutw.counts]);
-
-    match outct {
-        CarryType::Min => {
-            // Now, we must link the start node with the following node to whatever node the end
-            // node was connected.
-            ptgraph.graph.add_edge(startn,    outconn.0, EmptyEdge{t : outconn.1});
-            ptgraph.graph.add_edge(outconn.0, startn,    EmptyEdge{t : outconn.1.rev()});
-        },
-        CarryType::Max => {
-            // We must check what edges we had
-            let tmptype = EdgeType::from_carrytypes(CarryType::Min, outconn.1.get_from_and_to().1);
-            ptgraph.graph.add_edge(startn,    outconn.0, EmptyEdge{t : tmptype});
-            ptgraph.graph.add_edge(outconn.0, startn,    EmptyEdge{t : tmptype.rev()});
-        },
-    }
-    return true;
-}
-
-
-
 impl Shrinkable for PtGraph {
     type EdgeIdx = EdgeIndex;
     type NodeIdx = NodeIndex;
@@ -245,11 +42,12 @@ impl Shrinkable for PtGraph {
 
         let mut dididoanything = false;
         loop {
+            let mut dididoanythingnow = false;
             let ambnodes = self.get_ambiguous_nodes_bi(); // Just in case we hadn't got them yet
-            log::info!("Starting shrinking the graph with {} nodes and {} edges, beginning from {} ambiguous nodes",
+            logw(format!("Starting shrinking the graph with {} nodes and {} edges, beginning from {} ambiguous nodes",
                   self.graph.node_count(),
                   self.graph.edge_count(),
-                  ambnodes.len());
+                  ambnodes.len()).as_str(), Some("trace"));
             for an in ambnodes.iter() {
                 // println!("it");
                 // We need to see the forward edges from here
@@ -279,6 +77,7 @@ impl Shrinkable for PtGraph {
                         // println!("yes1");
                         self.shrink_single_path(*an, neigh[0].0, &ambnodes, neigh[0].1);
                         dididoanything = true;
+                        dididoanythingnow = true;
                     }
 
                 } else {
@@ -310,63 +109,16 @@ impl Shrinkable for PtGraph {
                                     // println!("yes2");
                                     self.shrink_single_path(n.0, outn[0].0, &ambnodes, outn[0].1);
                                     dididoanything = true;
+                                    dididoanythingnow = true;
                                 }
                             }
                         }
                     }
-
-                    // exit(1);
                 }
             }
-
-            // TEST ===== BUBBLE REMOVAL
-            log::info!("IMPORTANT: trivial resolution of standard bubbles is going to follow!");
-            let bubbles = self.graph.node_indices().filter(|n| {
-                    self.out_degree(*n) == 3
-                }).filter(|n| {
-                    // We can directly check only once the bubbles (and not do it twice) by fixing
-                    // the carrytype from which we want to continue
-                    let mut vmin = Vec::with_capacity(2);
-
-                    for e in self.graph.edges_directed(*n, Outgoing) {
-                        match e.weight().t.get_from_and_to().0 {
-                            CarryType::Min => {
-                                if vmin.len() == 2 {
-                                    return false;
-                                } else {
-                                    vmin.push(e.id());
-                                }
-                            },
-                            _ => {},
-                        }
-                    }
-
-                    // We have two outgoing from min/max and one outgoing from max/min.
-                    // We check whether this is actually a bubble or not.
-                    if vmin.len() == 2 {
-                        return check_bubble_structure(&self, *n, vmin);
-                    } else {
-                        return false;
-                    }
-                }).collect::<BTreeSet<NodeIndex>>();
-
-            if bubbles.is_empty() {
+            if !dididoanythingnow {
                 break;
-            } else {
-                logw(format!("Found {:?} potential bubbles (they might be less). Starting to collapse them ", bubbles.len()).as_str(), Some("trace"));
-                for n in bubbles {
-                    if self.graph.contains_node(n) {
-                        let tmpb = collapse_bubble(self, n);
-                        if tmpb {
-                            dididoanything = true;
-                        }
-                    }
-                }
             }
-            log::info!("IMPORTANT: trivial resolution of bubbles has been done.");
-            // TEST END ===== BUBBLE REMOVAL
-
-
         }
